+++ script/simulate.py
@@ -23,7 +23,7 @@ plt.rcParams['axes.unicode_minus'] = False
 # Global varibles
 c1 = c2 = c3 = c4 = c5 = c6 = c23 = c234 = 0
 s1 = s2 = s3 = s4 = s5 = s6 = s23 = s234 = 0
-d1, a3, a4, d4, d5, d6 = 89.459, 425, 392.25, 109.15, 94.65, 82.3
+d1, a2, a3, d4, d5, d6 = 89.459, 425, 392.25, 109.15, 94.65, 82.3
 R = np.zeros((3,3)); P = np.zeros((3,1)); Tran = np.zeros((4,4))
 jcb = ijcb = tjcb = np.zeros((6,6))
 
@@ -70,11 +70,11 @@ def ur_kinematics(q):
 
   ur_calcJnt(q)
   hori = d4 + d6*c5
-  vert = a3*c2 + a4*c23 - d5*s234 + d6*c234*s5
+  vert = a2*c2 + a3*c23 - d5*s234 + d6*c234*s5
 
   P[0][0] = -s1*hori + c1*vert
   P[1][0] = c1*hori + s1*vert
-  P[2][0] = d1 - a3*s2 - a4*s23 - d5*c234 - d6*s234*s5
+  P[2][0] = d1 - a2*s2 - a3*s23 - d5*c234 - d6*s234*s5
 
   R[0][0] = c6*(s1*s5 + c1*c234*c5) - s234*c1*s6
   R[0][1] = -s6*(s1*s5 + c234*c1*c5) - s234*c1*c6
@@ -93,7 +93,7 @@ def ur_kinematics(q):
 
 # 逆运动学
 def inv_kinematics(Tran, q):
-    global d1, a3, a4, d4, d5, d6
+    global d1, a2, a3, d4, d5, d6
     q_jnt = np.zeros(6)
     nx = Tran[0][0]; ox = Tran[0][1]; ax = Tran[0][2]; px = Tran[0][3]
     ny = Tran[1][0]; oy = Tran[1][1]; ay = Tran[1][2]; py = Tran[1][3]
@@ -115,21 +115,21 @@ def inv_kinematics(Tran, q):
     #  q_jnt[5] = math.atan2(-B6/s5,-A6/s5)
     s6 = math.sin(q_jnt[5]); c6 = math.cos(q_jnt[5])
     # 关节角3 (t14_14, t14_34)
-    A3 = px*c1 +py*s1 -d5*(nx*c1*s6+ox*c1*c6+ny*s1*s6+oy*s1*c6) -d6*(ay*s1+ax*c1)
+    a2 = px*c1 +py*s1 -d5*(nx*c1*s6+ox*c1*c6+ny*s1*s6+oy*s1*c6) -d6*(ay*s1+ax*c1)
     B3 = pz -d1 -az*d6 -d5*(oz*c6+nz*s6)
-    temp1 = math.acos((A3*A3+B3*B3-a3*a3-a4*a4)/(2*a3*a4))
-    temp2 = -math.acos((A3*A3+B3*B3-a3*a3-a4*a4)/(2*a3*a4))
+    temp1 = math.acos((a2*a2+B3*B3-a2*a2-a3*a3)/(2*a2*a3))
+    temp2 = -math.acos((a2*a2+B3*B3-a2*a2-a3*a3)/(2*a2*a3))
     q_jnt[2] = temp1 if abs(temp1-q[2])<abs(temp2-q[2]) else temp2
     s3 = math.sin(q_jnt[2]); c3 = math.cos(q_jnt[2])
     # 关节角2 (t14_14, t14_34)
-    A2 = -(a4*a4 +2*a3*a4*c3 +a3*a3)
-    B2 = -(a4*c3+a3)*A3 +a4*s3*B3
-    C2 = (a4*c3+a3)*B3 +a4*s3*A3
+    A2 = -(a3*a3 +2*a2*a3*c3 +a2*a2)
+    B2 = -(a3*c3+a2)*a2 +a3*s3*B3
+    C2 = (a3*c3+a2)*B3 +a3*s3*a2
     q_jnt[1] = math.atan2(C2/A2, B2/A2)
     # 关节角4 (t15_13, t15_33)
-    A4 = -(oz*c6 + nz*s6)
+    a3 = -(oz*c6 + nz*s6)
     B4 = -c6*(ox*c1+oy*s1) -s6*(nx*c1+ny*s1)
-    q_jnt[3] = math.atan2(B4, A4) - q_jnt[1] - q_jnt[2]
+    q_jnt[3] = math.atan2(B4, a3) - q_jnt[1] - q_jnt[2]
     if (q_jnt[3] > math.pi):
         q_jnt[3] = q_jnt[3] -2*math.pi
     elif (q_jnt[3] < -math.pi):
@@ -141,15 +141,15 @@ def ur_jacobian(q):
 
   ur_calcJnt(q)
 
-  jcb[:,[0]] = np.array([-(a3*c2+a4*c23-d5*s234+d6*c234*s5)*s1 - (d4+d6*c5)*c1,
-    -(d4+d6*c5)*s1+(a3*c2+a4*c23-d5*s234+d6*c234*s5)*c1, 0, 0, 0, 1]).reshape((6,1))
-  jcb[:,[1]] = np.array([-c1*(a3*s2+a4*s23+d5*c234+d6*s234*s5),
-    -s1*(a3*s2+a4*s23+d5*c234+d6*s234*s5),
-    -a3*c2-a4*c23+d5*s234-d6*c234*s5,
+  jcb[:,[0]] = np.array([-(a2*c2+a3*c23-d5*s234+d6*c234*s5)*s1 - (d4+d6*c5)*c1,
+    -(d4+d6*c5)*s1+(a2*c2+a3*c23-d5*s234+d6*c234*s5)*c1, 0, 0, 0, 1]).reshape((6,1))
+  jcb[:,[1]] = np.array([-c1*(a2*s2+a3*s23+d5*c234+d6*s234*s5),
+    -s1*(a2*s2+a3*s23+d5*c234+d6*s234*s5),
+    -a2*c2-a3*c23+d5*s234-d6*c234*s5,
     -s1, c1, 0]).reshape(6,1)
-  jcb[:,[2]] = np.array([-c1*(a4*s23+d5*c234+d6*s234*s5),
-    -s1*(a4*s23+d5*c234+d6*s234*s5),
-    -a4*c23+d5*s234-d6*c234*s5,
+  jcb[:,[2]] = np.array([-c1*(a3*s23+d5*c234+d6*s234*s5),
+    -s1*(a3*s23+d5*c234+d6*s234*s5),
+    -a3*c23+d5*s234-d6*c234*s5,
     -s1, c1, 0]).reshape((6,1))
   jcb[:,[3]] = np.array([-c1*(d5*c234+d6*s234*s5),
     -s1*(d5*c234+d6*s234*s5),
@@ -165,16 +165,22 @@ def ur_jacobian(q):
   ijcb = np.linalg.inv(jcb)
 
 
-dt = 10     # 采样间隔 (ms)
-T  = 20000  # 运行时间 (ms)
-w  = 30*d2r # 角速度 (rad/s)
+dt = 8     # 采样间隔 (ms)
+T  = 500  # 运行时间 (ms)
+w  = 40*d2r # 角速度 (rad/s)
 v  = 0      # 线速度 (mm/s)
 nn = int(T/dt+1) # 采样点数
 q0 = np.array([0, -90, 90, -90, -90, 0]).reshape((6,1))*d2r  # 初始关节角
+#  q0 = np.array([0, 0, 0, 0, -90, 90]).reshape((6,1))*d2r  # 初始关节角
 pos = np.zeros((3,nn))
 q_jnt = np.zeros((6,nn))
 dx = np.array([v/(T/dt), 0, 0, w/(T/dt), 0, 0]).reshape((6,1))  # 每个周期的状态变化
 time = np.linspace(0, int(T), int(T/dt+1)) # 时间
+
+ur_kinematics(q0)
+#  pos[:,0] = Tran[0:3,3]
+ur_jacobian(q0)
+print(ijcb)
 
 # 循环开始
 #  q = q0
@@ -189,10 +195,10 @@ time = np.linspace(0, int(T), int(T/dt+1)) # 时间
 
 # 初始状态
 #  q_jnt[:,0] = q0.reshape(1,6)
-q0 = [0, -98.9, 117.79, -108.88]
-ur_kinematics(q0)
+#  q0 = [0, -98.9, 117.79, -108.88]
+#  ur_kinematics(q)
 #  pos[:,0] = Tran[0:3,3]
-print(Tran)
+#  print(Tran)
 # 终止状态
 #  Tran = [[1,0,0,300],[0,1,0,300],[0,0,1,300],[0,0,0,1]]
 #  qq = inv_kinematics(Tran, q0)
@@ -317,6 +323,6 @@ ax3.legend(loc='best')
 # make sure that the plots fit nicely in your figure
 plt.tight_layout()
 #  plt.savefig('../foo.svg')
-plt.show()
+#  plt.show()