+++ src/trajectory.cpp
@@ -11,7 +11,7 @@ void calc_ref_joint(SVO& svo) {
   // 当前路径执行的时间
   double offsetTime = svo.time - svo.path.beginTime;
   double freq = svo.path.freq;
-  double delQ = 1.4*Deg2Rad;
+  double delQ = 8*Deg2Rad;
 
   // 角度伺服
   if (svo.path.angleServo) {
@@ -46,21 +46,25 @@ void calc_ref_joint(SVO& svo) {
  */
 bool velocity_interpolation(double offsetTime, double freq, THETA curTheta, 
     THETA& refTheta, ARRAY velocity) {
-  // 当前运动时间占总时间的比例大于1，即超过预定时间，插值点不变，返回已完成
-  if (freq * offsetTime > 1) return true;
+  Vec6d velo, dq;
+  double time = freq * (offsetTime+SERVO_TIME);
 
-  JACOBIAN* jcbCompact = new JACOBIAN;
-  MATRIX_D jcb = Zeros(6,6);
-  MATRIX_D velo = Zeros(6,1), dq = Zeros(6,1);
-
+  for (int i=0; i<6; ++i) velo[i] = velocity[i];
   // 用参考角度计算各关节角的正余弦值
-  calcJnt(curTheta);
-  for (int i=0; i<6; ++i) velo(i+1,1) = velocity[i];
+  calcJnt(refTheta);
   // 计算 Jacobian 矩阵
-  jcb = ur_jacobian(jcbCompact);
-  MATRIX_D ijcb(6,6, (double *)jcbCompact->invJcb);
+  Mat6d ijcb = ur_jacobian().inverse();
   dq = ijcb*velo;
-  for (int i=0; i<6; ++i) refTheta[i] += dq(i+1,1);
+  // 当前运动时间占总时间的比例大于1，即超过预定时间，插值点不变，返回已完成
+  if (time >= 1) {
+    // 路径停止前，给一定时间让机械臂停止
+    dq = Vec6d(0);
+    for (int i=0; i<6; ++i) refTheta[i] = refTheta[i];
+    if (time >= 1.1) {
+      return true;
+    }
+  }
+  for (int i=0; i<6; ++i) refTheta[i] += dq(i);
   return false;
 } // bool velocity_interpolation()
 
@@ -74,10 +78,10 @@ bool velocity_interpolation(double offsetTime, double freq, THETA curTheta,
 bool joint_interpolation(double offsetTime, double freq, int interpMode, 
     ARRAY orig, ARRAY goal, ARRAY& refVal) {
   // 当前运动时间占总时间的比例 | 已完成的路径占全路径的比例
-  double time = freq * offsetTime;
+  double time = freq * (offsetTime+SERVO_TIME);
 
   // 超过预定时间，插值点设为终点值，返回已完成
-  if (time > 1) {
+  if (time >= 1) {
     refVal = goal;
     return true;
   }